COMP: Changed int -> double to avoid warnings.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=1585
git-svn-id: http://svn.slicer.org/Slicer4/trunk@1585 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLCameraNode.h
@@ -96,14 +96,14 @@ public:
 
   // Description:
   // Set camera Parallel Scale 
-  void SetParallelScale(int scale) 
+  void SetParallelScale(double scale) 
     {
     this->Camera->SetParallelScale(scale);
     };
   
   // Description:
   // Set camera Parallel Scale   
-  int GetParallelScale()
+  double GetParallelScale()
     {
     return this->Camera->GetParallelScale();
     };
